Update README.md
#### Hi👋 I'm teenager-ETH
![Hi👋 I'm teenager-ETH](https://arturssmirnovs.github.io/github-profile-readme-generator/images/banner.png)

Web3 enthusiast and blockchain technology specialist with experience in developing decentralized applications and DeFi solutions. Currently starting my programming journey by learning the fundamentals of coding and smart contract development. Passionate about creating innovative solutions to automate processes and expand the potential of blockchain technology. Continuously learning and experimenting to grow as a developer and contribute to the advancement of the Web3 ecosystem.

- 🔭 I’m currently working on this page. 


[<img src='https://cdn.jsdelivr.net/npm/simple-icons@3.0.1/icons/github.svg' alt='github' height='40'>](https://github.com/teenager-ETH)  [<img src='https://cdn.jsdelivr.net/npm/simple-icons@3.0.1/icons/twitter.svg' alt='twitter' height='40'>](https://twitter.com/https://x.com/Maksym010101)  

![GitHub stats](https://github-readme-stats.vercel.app/api?username=teenager-ETH&show_icons=true)